initialize interface , adding anchor panes
Home class now implements initializable interface which helps us to fetch the post when the home page loaded . Initializable interface consists of a method called initalize. which called the method to populate the posts. populateposts method fetch the posts from database and calls a create panes method to create  a pane for individual post. then it adds it in a postContainer to display. The createPane method initialize the new AnchorPane then adds the required labels and then add the button to apply . All these are added to AnchorPane and AnchorPane is added to postContainer to be displayed